Correccion de datos en la CapaDatos

+++ code/CapaEntidades/MediosPago.cs
@@ -10,6 +10,6 @@ namespace CapaEntidades
     {
         public int id_medioPago { get; set; }
         public string nombre_medio { get; set; }
-        public float comision { get; set; }
+        public decimal comision { get; set; }
     }
 }
